+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * 【core  】      修复TypeUtil.getTypeArgument对实现接口获取不全面问题（issue#I7CRIW@Gitee）
 * 【core  】      修复BeanUtil.isCommonFieldsEqual判空导致的问题
 * 【extra 】      修复CompressUtil.createArchiver 将文件压缩为tgz时文件名规则无效问题（issue#I7LLL7@Gitee）
+* 【core  】      修复脱敏银行卡号长度bug（pr#3210@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.20(2023-06-16)