+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # 🚀Changelog
 
 -------------------------------------------------------------------------------------------------------------
-# 5.8.21(2023-07-19)
+# 5.8.21(2023-07-20)
 
 ### 🐣新特性
 * 【core  】      list 为空时，CollUtil.max等返回null而非异常（pr#1027@Gitee）
@@ -12,7 +12,8 @@
 * 【core  】      RandomUtil增加可选是否包含边界的重载（issue#3182@Github）
 * 【core  】      StrUtil增加truncateByByteLength方法（pr#3176@Github）
 * 【core  】      身份证工具类isValidCard18、isValidCard15入参null直接返回null（pr#1034@Gitee）
-* 【http  】      使用multiparty方式支持body参数（issue#3158@Gitee）
+* 【http  】      使用multiparty方式支持body参数（issue#3158@Github）
+* 【core  】      ZipReader增加setMaxSizeDiff方法，自定义或关闭ZipBomb（issue#3018@Github）
 
 ### 🐞Bug修复
 * 【core  】      修复MapUtil工具使用filter方法构造传入参数结果问题（issue#3162@Github）